Add license header for each file and each manager provides method to persist data

+++ Test/Doctrine/EntityManager.php
@@ -1,4 +1,12 @@
 <?php
+/**
+ * This file is part of the Oxygen Bundle Package.
+ *
+ * (c) 2014 Oxyfony
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 namespace O2\Bundle\ModelBundle\Test\Doctrine;
 
 use Doctrine\ORM\EntityManager as DoctrineEntityManager;
@@ -28,9 +36,13 @@ trait EntityManager {
 	 */
 	protected function getEntityManager()
 	{
-		if (!is_null($this->em))
-			return $this->em;
+		if (is_null($this->em))
+			$this->em = $this->createEntityManager();
 		
+		return $this->em;
+	}
+	
+	protected function createEntityManager($overrideDatabase = true) {
 		// Bundle namespace
 		$bundleNamespace = explode('\\', __NAMESPACE__);
 		array_pop($bundleNamespace); array_pop($bundleNamespace);
@@ -80,8 +92,10 @@ trait EntityManager {
 		$schemaTool = new SchemaTool($em);
 		$cmf = $em->getMetadataFactory();
 		$classes = $cmf->getAllMetadata();
-		$schemaTool->dropDatabase();
-		$schemaTool->createSchema($classes);
+		if ($overrideDatabase) {
+			$schemaTool->dropDatabase();
+			$schemaTool->createSchema($classes);
+		}
 	
 		return $em;
 	}
